requirements.txt
certifi==2018.11.29
chardet==3.0.4
Cython==0.29.2
docutils==0.14
idna==2.8
Kivy-Garden==0.1.4
pip==10.0.1
Pygments==2.3.1
requests==2.21.0
setuptools==39.1.0
urllib3==1.24.1
